+++ stream-chat-android-compose-sample/src/main/java/io/getstream/chat/android/compose/sample/ChatApp.kt
@@ -8,7 +8,6 @@ import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.notifications.handler.NotificationConfig
 import io.getstream.chat.android.client.notifications.handler.NotificationHandler
 import io.getstream.chat.android.client.notifications.handler.NotificationHandlerFactory
-import io.getstream.chat.android.client.utils.internal.toggle.ToggleService
 import io.getstream.chat.android.core.ExperimentalStreamChatApi
 import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.offline.experimental.plugin.Config
@@ -28,13 +27,7 @@ class ChatApp : Application() {
         dateFormatter = DateFormatter.from(this)
 
         setupStreamSdk()
-        initializeToggleService()
         connectUser()
-    }
-
-    @OptIn(InternalStreamChatApi::class)
-    private fun initializeToggleService() {
-        ToggleService.init(applicationContext, mapOf(ToggleService.TOGGLE_KEY_OFFLINE to BuildConfig.DEBUG))
     }
 
     @OptIn(ExperimentalStreamChatApi::class)