+++ stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/viewmodel/messages/MessageListViewModel.kt
@@ -421,7 +421,8 @@ public class MessageListViewModel @JvmOverloads constructor(
                                 attachment.imageUrl == attachmentToBeDeleted.imageUrl
                             }
                         }
-                        domain.editMessage(message).enqueue(
+
+                        client.updateMessage(message).enqueue(
                             onError = { chatError ->
                                 logger.logE("Could not edit message to remove its attachments: ${chatError.message}. Cause: ${chatError.cause?.message}")
                             }